feat: move helper to utils
