+++ client_src/BaseMenu/moretab.js
@@ -16,7 +16,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { translate } from '../localizer';
-import { write } from '../helper';
+import { write } from '../utils/helper';
 
 // Render "More" page.
 global.rMore = function () {
