+++ client_src/utils/helper.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { translate } from './localizer';
+import { translate } from '../localizer';
 
 declare const sectorWidth: number;
 declare const bigNotes: any[];
